+++ robot_descriptions/_descriptions.py
@@ -93,6 +93,7 @@ DESCRIPTIONS: Dict[str, Description] = {
     "gen2_description": Description(Format.URDF),
     "ginger_description": Description(Format.URDF),
     "go1_description": Description(Format.URDF),
+    "go1_mj_description": Description(Format.MJCF),
     "hyq_description": Description(Format.URDF),
     "icub_description": Description(Format.URDF),
     "iiwa_description": Description(Format.URDF),