+++ robot_descriptions/go1_mj_description.py
@@ -16,17 +16,19 @@
 # limitations under the License.
 
 """
-Go1 description.
+Go1 MJCF description.
 """
 
 from os import path as _path
 
 from ._cache import clone_to_cache as _clone_to_cache
 
-REPOSITORY_PATH: str = _clone_to_cache("unitree_ros")
+REPOSITORY_PATH: str = _clone_to_cache("unitree_mujoco")
 
-PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "robots", "go1_description")
+PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "data", "go1")
 
 MESHES_PATH: str = _path.join(PACKAGE_PATH, "meshes")
+
+MJCF_PATH: str = _path.join(PACKAGE_PATH, "xml", "go1.xml")
 
 URDF_PATH: str = _path.join(PACKAGE_PATH, "urdf", "go1.urdf")