+++ robot_descriptions/allegro_hand_description.py
@@ -29,8 +29,6 @@ PACKAGE_PATH: str = _path.join(
     REPOSITORY_PATH, "manipulation", "models", "allegro_hand_description"
 )
 
-MESHES_PATH: str = _path.join(PACKAGE_PATH, "meshes")
-
 URDF_PATH: str = _path.join(
     PACKAGE_PATH, "urdf", "allegro_hand_description_left.urdf"
 )