+++ robot_descriptions/tiago_description.py
@@ -27,8 +27,6 @@ REPOSITORY_PATH: str = _clone_to_cache("example-robot-data")
 
 PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "robots", "tiago_description")
 
-MESHES_PATH: str = _path.join(PACKAGE_PATH, "meshes")
-
 URDF_PATH: str = _path.join(PACKAGE_PATH, "robots", "tiago.urdf")
 
 # Description-specific paths