+++ robot_descriptions/__init__.py
@@ -19,4 +19,4 @@
 Import open source robot description as Python modules.
 """
 
-__version__ = "0.4.0rc1"
+__version__ = "0.4.0"